Module Cow

Camel on the Web (CoW) library.

This library includes module aliases for various useful Web formats such as XML, XHTML, JSON, Markdown and Atom. In many cases, you can also use the constitutent libraries directly.

module Xml : sig ... end

XML combinators.

module Json : sig ... end

JSON (JavaScript Object Notation) library

module Markdown : sig ... end

Markdown library

module Html : sig ... end

(X)HTML library

module Xhtml : sig ... end

XHTML entity support

module Atom : sig ... end

The Atom Syndication format. See RFC4287 for the full specification.