Module Cow.Markdown

Markdown library

type t = Html.t

The type of a Markdown document, represented as an (X)HTML tree

val to_string : t -> string

to_string t will output an (X)HTML string representation of the t Markdown

val of_string : string -> t

of_string s will parse the s Markdown and store an X(HTML) representation via the Omd library.